requirements.txt
annotated-types==0.7.0
anyio==4.10.0
audioop-lts==0.2.2
av==15.1.0
certifi==2025.7.14
charset-normalizer==3.4.3
click==8.1.8
coloredlogs==15.0.1
ctranslate2==4.6.0
faster-whisper==1.2.0
filelock==3.20.0
flatbuffers==25.9.23
fsspec==2025.9.0
greenlet==3.2.3
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
humanfriendly==10.0
idna==3.10
mpmath==1.3.0
numpy==2.3.3
ollama==0.6.0
onnxruntime==1.23.1
packaging==25.0
psutil==6.1.0
protobuf==6.32.1
pydantic==2.12.0
pydantic_core==2.41.1
pydub==0.25.1
python-dotenv==1.1.1
python-telegram-bot==22.5
PyYAML==6.0.3
requests==2.32.5
schedule==1.2.2
setuptools==80.9.0
sniffio==1.3.1
SQLAlchemy==2.0.43
aiosqlite==0.20.0
sympy==1.14.0
tokenizers==0.22.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.14.1
urllib3==2.5.0
